fix user_id type

+++ atoma-state/src/migrations/20250424111320_fiat_payments.sql
@@ -1,6 +1,6 @@
 CREATE TABLE
   IF NOT EXISTS fiat_balance (
-    user_id INTEGER NOT NULL PRIMARY KEY,
+    user_id BIGINT NOT NULL PRIMARY KEY,
     usd_balance BIGINT NOT NULL DEFAULT 0,
     already_debited_amount BIGINT NOT NULL DEFAULT 0,
     overcharged_unsettled_amount BIGINT NOT NULL DEFAULT 0,
@@ -9,7 +9,7 @@ CREATE TABLE
 
 CREATE TABLE
   IF NOT EXISTS usage_per_model (
-    user_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
     model TEXT NOT NULL,
     total_number_processed_tokens BIGINT NOT NULL DEFAULT 0,
     PRIMARY KEY (user_id, model)
